pyproject.toml
[project]
name = "pftl"
version = "0.1.0"
description = "Space ship battle simulation game. Two to four player matches."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.4",
    "sqlalchemy>=2.0.30",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
]
